Move AuctionStatus models to ViewModel module
